QDoc: Remove unnecessary link warnings
When QDoc handles an example it generates a series of links for it, such
as a link of files that are present in the example.

Internally, those links are represented as `Atom`s, an internal
representation for most elements that compose a documentation block.

In particular, they are represented as `Atom`s of types
`ExampleFileLink` and `ExampleImageLink`.

When an `ExampleFileLink` or `ExampleImageLink` is handled by QDoc, when
generating the output documentation, a link is built and, if it fails to
link, a warning is issued to notify the user of the failed linking.

Nonetheless, those kind of links are built internally. QDoc finds and
handles a series of files and images related to an example and generates
a relevant `Atom` for linking only when those files or images are
resolved.

That is, those links are expected to be correct by construction, so that
failure to link them is to be related to an internal QDoc bug rather
than something the user did.

Hence, reporting the linking failure to the user is unnecessary, as the
user doesn't have the required control to fix the issue.

Thus, QDoc will now avoid producing link warnings when failing to link
to an example image or file.

Change-Id: Ifbbc92d0e32bf38270c3f9aec2ba2c2a6fc81f09
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>