QDoc: Report the location of links that fail to be linked
QDoc allows the user to cross-reference a documentation set or an
external resource by the use of the "\l" command.

For example, a documentation set might have a page titled `Foo` and a page
titled `Bar` that are related, so that the user has a desire to link
from `Bar` to `Foo` to allow a reader to pick up some relevant
information.

The user can do so by using an "\l" command, such as "\l Foo" in the
documentation for the `Bar` page.

When QDoc produces the output documentation, it tries to resolve each
reference that a user provided by the usage of the "\l" command.

When QDoc fails to understand what the user is referring to, it issues a
warning to notify the user of the failed linkage.

The produced warning contains some information about the position of the
failing link command. Nonetheless, due to some internal architectural
issues, the reported position is always that of the documentation block
that contains the link, with no information about where the link is
positioned in the block.

For example, a block starting at line X in file `Boo`, that contains a
series of links at various positions in its body, will report all
failing links as being in file `Boo` at position X, making it difficult
for the user to understand where they should be fixing the warning, with
the difficulty growing by the size of the documentation block and the
amount of links in it.

To simplify recognizing which link command is failing, QDoc will now
report more precise line information for failing links.

When QDoc parses an "\l" command, it produces a `LinkAtom`, an internal
representation that is used to store information about a user provided
reference.

Later on, when QDoc is producing the final output documentation, it will
try to resolve the `LinkAtom`, reporting a warning if it fails to do so.

A `LinkAtom` does not have any information about where its generating
command is located, as that information is only kept by QDoc during
parsing and discarded after.

In turn, this makes it impossible for QDoc to provide correct location
information when the `LinkAtom` fails to resolve when generating the
output documentation.

To avoid the issue, a `LinkAtom` now stores a `Location`, the internal
representation for source file positions, that be used to
report more precise information when failing to resolve the link.

`LinkAtom` was enhanced with a new `location` member.
The various constructors for `LinkAtom` were modified to ensure the
initialization of the new member.

The main constructor for `LinkAtom` was modified to allow the user to
pass in an optional `Location`, that will be stored in the new
`location` member.

In `DocParser`, the class taking care of parsing a user provided
documentation block, QDoc constructs a `LinkAtom` when encountering an
"\l" command.

`DocParser`, which keeps track of the current location while parsing,
will now provide an adequately precise `Location` when constructing a
`LinkAtom`.

The issuing of a failed link warning in `HtmlGenerator`, which takes
care of producing the final output documentation for the HTML format,
will now access the stored `location` member in a `LinkAtom`, when
available, to use when reporting the failed link warning, instead of
using the `Location` of the documentation block that contains the
non-resolved link.

Fixes: QTBUG-119999
Change-Id: I0a83ffa98f8012496ec739277314ec1852eee9f8
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>