QDoc: Add output validation test
A test called tst_generatedOutput already exists for testing QDoc's
generated files. That test suffers from multiple shortcomings, such as
hand-rolled comparison, a joint expected output directory for all
tests, and explicit listing of expected files, to name a few. These
properties of tst_generatedOutput make it cumbersome to maintain and
extend that test.

As the material form of the output test is changing, comparing test
results against tst_generatedOutput as reference may be misleading for
the sake of proper test statistics. Therefore, this patch introduces a
new test for QDoc, tst_validateQdocOutputFiles. It runs QDoc in order
to generate documentation for projects contained in its `testdata`
subdirectory. The test is data driven, populated by the QDoc projects
(identified by .qdocconf files) found in the subdirectories of
`testdata`. The implementation is partially taken from
tst_generatedOutput.

The intention is to let this new test replace tst_generatedOutput as
soon as coverage parity is achieved.

This patch includes:
- The test itself.
- A `testdata` directory that contains a simple test that replicates
  the minimum QDoc configuration according to the QDoc manual, and its
  expected output. This project serves two purposes: it validates that
  the test works as intended, and it keeps the QFETCH macro from
  asserting over missing testdata.
- Necessary updates to build system files.
- A `README.md` for the test that describes its intended purpose and
  how to extend it.

Task-number: QTBUG-119500
Change-Id: I8e8295d2040fa039d66d36861138889ef32d30af
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>