QDoc: Remove topic argument from processTopicArgs
When QDoc parses a user-provided documentation block it converts it into
a `Doc`, an intermediate representation that is used internally.

A documentation block is parsed into a series of elements, some of which
are "topic commands".

Topic commands are commands that are used to identify which documentable
element the documentation block refers to.

Generally, topic commands are then "processed" by
`CppCodeParser::processTopicArgs`, which takes care of understanding
what element the topic specifically refers to.

`CppCodeParser::processTopicArgs` takes, among others, a string
representing the topic command name to be process and a `Doc`
representation of the comment block containing the topic command.

Nonetheless, a `Doc` representation already has knowledge about the
topic commands that were used in the documentation block.

Indeed, the consumers of `CppCodeParser::processTopicArgs` generally
extract the first topic command stored in the `Doc` representation and
pass it along with the `Doc` representation itself.

Since it is possible to easily extract the topic command from the `Doc`
representation, the topic argument was removed from
`CppCodeParser::processTopicArgs`, in favor of retrieving the topic from
the already passed in `Doc` representation.

`CppCodeParser::processTopicArgs` was modified in both declaration and
definition to remove the previously present `topic` argument.

The implementation of `CppCodeParser::processTopicArgs` was modified to
extract the "main topic", which the `topic` argument previously
represented, directly from the passed in `Doc`.

As a `Doc` may not have any topic command, albeit this will not
currently be the case with how the code is structured on the consumers
side, an early return was added to `CppCodeParser::processTopicArgs` in
case no topic command is present, as there is nothing the method should
do in that case.

The consumers of `CppCodeParser::processTopicArgs` were modified to stop
passing the `topic` argument when calling the method.

Change-Id: Ic360f79584c7f2486639b5a7c7a617337ceaf9d6
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>