QDoc: Remove out parameters for processTopicArgs
A user provided documentation block can contain zero or more topic
commands.

A topic command is used to identify which documentable element the
documentation block refers to.

When QDoc parses a user provided documentation block, it registers which
topic commands, if any, are present, in a `Doc`, the internal
representation for documentation blocks, later resolving them to various
`Nodes`, the internal representation for documentable elements.

`CppCodeParser::processTopicArgs` performs this process.

`processTopicArgs` accepts two out parameters, a list of `Node`s and a
list of `Doc`s.

When `processTopicArgs` is used, it obtains, as part of its processing,
a series of `Node`s that represent the documentable element that the
process topic refers to.

It stores the found `Node`s into the output parameter, together with the
`Doc` that the topic was produced from, the one the found `Node`s are
"tied" to.

The consumers of `processTopicArgs` later use the modified output
parameters to process the rest of the `Doc` commands, generally in
`CppCodeParser::processMetaCommands`, passing them along so that further
processing can be performed on the "tied" `Node`s/`Doc`s pairs.

While output parameters are used, all consumers of `processTopicArgs`
provide empty lists as the output parameters. That is, the contents of
the output parameters are only decided by `processTopicArgs` itself.

Hence, to simplify reasoning about the `Node`s and `Doc`s lists produced
in `processTopicArgs` and to reduce the scope of the mutable state that
the output parameters represent, the output parameters are now removed
in favor of returning a value from `processTopicArgs`.

The definition and declaration of `processTopicArgs` were modified to
remove the two parameters. Furthermore, the return type was modified to
be a pair of

The definition of `processTopicArgs` was modified to instantiate the two
lists, instead of obtaining them as arguments, further returning them at
the end of the computation.

The consumers of `processTopicArgs` were modified to stop passing the
two out parameters when calling the method, instead assigning its
returned value to obtain the two lists.

Change-Id: Id987f45533cc9428d58c23ac80d842e7afcb8ff9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>