QDoc: Ensure `nosubdirs` doesn't mangle `outputdir`
QDoc documentation projects must define an `outputdir` that specifies
the directory into which QDoc will create files for the project. This
variable can be defined in the project's .qdocconf file, or with the
`-outputdir` command line option.

QDoc can generate output in various formats, specified by the
`outputformats` configuration directive. Each output format can specify
a directory into which QDoc shall place the files created for that
format through the `outputsubdir` configuration variable.

If the configuration for a project specifies the `nosubdirs` variable
for any of the configured output formats, QDoc drops the last directory
in the path hierarchy of the output directory (specified with the
`outputdir` configuration variable or passed via `-outputdir` on the
command line) for that format. Thus, given the following configuration:

    outputdir = /tmp/foo
    outputformats = HTML DocBook
    HTML.outputsubdir = html
    HTML.nosubdirs = true
    DocBook.outputsubdir = docbook
    DocBook.nosubdirs = true

Instead of placing output in `/tmp/foo/html` and `/tmp/foo/docbook`,
QDoc will place its output into `/tmp/html` and `/tmp/docbook`.
`/tmp/foo` will neither be created nor (if pre-existing) modified.

If `nosubdirs` is set and the `outputsubdir` variable isn't specified
for the given format, QDoc defaults to the name `html`.

This change modifies QDoc's behavior such that, when both `nosubdirs`
and `outputsubdir` are set, the last directory in the `outputdir`
hierarchy isn't removed before appending `outputsubdir`. This ensures
that the path specified by `outputdir` will be the top level directory
unless the `outputsubdir` explicitly navigates out of that hierarchy
with a relative path.

[ChangeLog][QDoc][Behavior Change] The `nosubdirs` configuration
variable no longer removes the last segment in the `outputdir` path.
To retain previous behavior for projects that define `outputsubdir`
and `nosubdirs`, prepend `outputsubdir` with `../`.

Task-number: QTBUG-77650
Task-number: QTBUG-119500
Change-Id: I310b86db3689d6236cd6dc213b1dbd28906f6531
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>