QDoc: Use getDefaultArgs in parseFnArg
`ClangCodeParser::parseFnArg` is used to have Clang parse the argument
of an "\fn" command to produce an AST that is used to find out which
element the writer of the command referred to.

When Clang is used to parse the argument of the "\fn" command, it is
necessary to pass a series of arguments for the compilation that Clang
needs to perform.

To do so, a vector is used and filled with a series of pre-existing
values that are generally used when dispatching a compilation to Clang.

The code that initializes this vector, performs the same computation
that `ClangCodeParser::getDefaultArgs` does.
The only difference is that `getDefaultArgs` uses an internal vector
that is persistent between calls and cleared when `getDefaultArgs` is
called, while `parseFnArg` uses an scope-limited vector.

For the purposes of the lifetime of ClangCodeParser and its current
usage, the internal argument vector that is used by `getDefaultArgs` has
no conflict usage conflicts between calls to `parseFnArg`.

Hence, to reduce code duplication, the scope-limited vector in
`parseFnArg` is removed in favor of using the internal argument vector
and the code that performs the same computation as
`ClangCodeParser::getDefaultArgs` is substituted by a call to `getDefaultArgs`.

Change-Id: I7156ad41a54b0f4a8d81e4c04ab060aac95e763c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>