QDoc: Remove ClangCodeParser::precompileHeaders
When QDoc parses a project that has C++ code as part of its sources, it
precompiles all of the project headers to produce a PCH that is later
used when parsing the actual implementation files, as a form of
optimization.

`ClangCodeParser`, the entity taking care of interfacing with Clang,
exposes `precompileHeaders`, the entry point to building the required
PCH.

`precompileHeaders` initializes some required internal state and then
dispatches to a private method, `buildPCH`, that performs the actual
compilation.

To reduce the code size and make mutable state mutation more local to
their actual scope, `precompileHeaders` is now removed and inlined in
`buildPCH`.

`buildPCH` was made public as a replacement for `precompileHeaders` and
the single call to `precompileHeaders` in "main.cpp" was substituted
with a call to `buildPCH`.

As part of the state initialization, `precompileHeaders` would build a
`CXIndex`, as required by Clang's APIs, that needs to be disposed
manually.

Due to `buildPCH` having multiple return points, the disposal was
duplicated to ensure that it is performed at each exit point of
`buildPCH`.

Change-Id: Ie6bbda4e77d5da4a626808fa0d9ea923d86915ed
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>