QDoc: Remove static CXIndex in clagcodeparser.cpp
QDoc utilizes libclang, Clang's C API, to parse the source code of
projects that makes use of C++ code.

libclang exposes a `CXIndex` type, that is used when dispatching a
compilation to it.

`clangcodeparser.cpp`, which dispatches various compilation tasks to
libclang, creates some `CXIndex`s as part of its code.

`clangcodeparser.cpp` provides a static `CXIndex` variable that is used
every time a `CXIndex` is obtained from libclang.

The create `CXIndex`s in "clangcodeparser.cpp" are always created and
disposed as part of the same scope, so that the static scope is
unnecessarily broad compared to the actual lifetime of its values.

Based on other similar patterns in QDoc, this is intended as a
microoptimization.

Nonetheless, the broad lifetime makes it more difficult to reason about
the code, prevents certain future improvements such as parallelization
efforts for the computations in "clangcodeparser.cpp" and is not proved
to be a meaningful optmization effort based on the performed profilings.

Hence, the static `CXIndex` variable in "clangcodeparser.cpp" is now
removed in favor of scope-local `CXIndex` variables.

Change-Id: I2994d63e7a2b6e103ee7b68d2473e3465b4edb31
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>