QDoc: Automate disposal of `CXIndex`s
QDoc uses libclang, Clang's C API, to obtain information about the
C++ source code in a project.

As part of the normal usage of the APIs, libclang exposes `CXIndex`
which is generally used to store information about a series of
translation units.

A `CXIndex` is a pointer and needs to be disposed manually by the usage
of `clang_disposeIndex`.

As usual, this may be error prone, especially for those cases where
there are multiple return paths, such as some of the `CXIndex` using
computations in "clangcodeparser.cpp", the implementation file where all
usages of libclang are encapsulated.

Hence, a new record, `CompilationIndex` is now provided in
"clangcodeparser.cpp".

`CompilationIndex` is a wrapper for `CXIndex` that treats the contained
`CXIndex` as a resource and disposes it on destruction.

A conversion operator to `CXIndex` was provided for `CompilationIndex`
to reduce the surface of the changed code and avoid additional noise.

Each usage of `CXIndex` is now replaced by `CompilationIndex`.

Manual calls to `clang_disposeIndex` were removed as they are now unnecessary.

Change-Id: I611b28ad552b51af98ea68212289a4fc17123326
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>