QDoc: Automate disposal of `CXTranslationUnit`s
QDoc uses libclang, Clang's C API, to obtain information about the
C++ source code in a project.

As part of the normal usage of the APIs, libclang exposes
`CXTranslationUnit` which is generally used to store information about a
specific translation unit and access the AST that Clang built.

A `CXTranslationUnit` is a pointer and needs to be disposed manually by
the usage of `clang_disposeTranslationUnit`.

As usual, this may be error prone, especially for those cases where
there are multiple return paths, such as some of the `CXTranslationUnit`
using computations in "clangcodeparser.cpp", the implementation file
where all usages of libclang are encapsulated.

Hence, a new record, `TranslationUnit` is now provided in
"clangcodeparser.cpp".

`TranslationUnit` is a wrapper for `CXTranslationUnit` that treats the
contained `CXTranslationUnit` as a resource and disposes it on
destruction.

A conversion operator to `CXTranslationUnit` was provided for
`TranslationUnit` to reduce the surface of the changed code and avoid
additional noise.

Furthermore, a conversion operator to `bool` was provided.

Normally, a `CXTranslationUnit` is initialized to a `nullptr` and a
concrete value is obtained by the usage of `clang_parseTranslationUnit /
clang_parseTranslationUnit2`.

The obtained `CXTranslationUnit` is then checked as a boolean value to
ensure that the parsing operation was successful.

This usage is covered by the implicit conversion to `bool`, to avoid
touching additional code.

The usage of `clang_parseTranslationUnit2`, which is favored over
`clang_parseTranslationUnit` and is used over it in QDoc, expects a
`CXTranslationUnit` to be passed as an out-parameter, which is later
filled if the operation is successful.

While this case could be covered by an implicit conversion to
`CXTranslationUnit*`, this would require the removal of the address
operator from the usage site, which would in turn make the fact that we
are dealing with an out-parameter less visible.

Hence, the implicit conversion for this case was not implemented, in
favor of modifying the calls to `clang_parseTranslationUnit2` to
directly pass the `tu` member of a `TranslationUnit` instance.

Each usage of `CXTranslationUnit` is now replaced by `TranslationUnit`.

Manual calls to `clang_disposeTranslationUnit` were removed as they are
now unnecessary.

Change-Id: I7f55686c92dacb334b2fe3fe65d4d6d2995deb39
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>