QDoc: Minor fixes in the QDoc manual
The QDoc manual contains incorrect syntax for the use of QDoc's
`extraimages` configuration variable. This change corrects the mistake.
As "drive-by" change, the surrounding text is modified to make it
clearer.

Inspecting the documentation for the `stylesheets` and `scripts`
configuration variables (they're mentioned in the bug report) exposed
another oddity in the manual, which is also addressed herein.

Pick-to: 6.6 6.5
Fixes: QTBUG-121226
Change-Id: I94e37b29de20dc932a2de532d6e5a4f4a7d36512
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 55bf738c007582ce771bc10ce45298f6d68375f9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>