qdoc: Use the correct output subdirectory when resolving dependencies
After commit c51980bb, Config::getOutputDir() now returns a clean path
constructed from -outputdir command line argument (or 'outputdir'
configuration variable), and a format-specific output subdirectory
if one is defined in the configuration file.

This also allowed the use of relative paths as the subdirectory, e.g.

    HTML.outputsubdir = ../html

which is used when generating online Qt documentation where output from
all modules' docs are directed to a shared output directory.

In main.cpp, the code that locates and loads .index files was using
code that duplicated parts of the now-changed logic for constructing
output directories. Update that logic to use Config::getOutputDir()
instead, to get name of the output subdirectory.

The subdirectory name, without path, is the only information that's
needed for searching .index files, because that subdirectory is
expected to exist under one of the locations provided with
-indexdir command line argument(s).

Task-number: QTBUG-77650
Change-Id: Idf8f23695873a260148defb3780a855e6df71a00
Reviewed-by: Paul Wicking <paul.wicking@qt.io>