QDoc: Make ClangCodeParser::printDiagnostics a free function
`ClangCodeParser::printDiagnostics` does not depend or use any of
`ClangCodeParser` state and is only used as part of the code in
"clangcodeparser.cpp".

Hence, the method was removed in favor of a static free function in
"clangcodeparser.cpp".

The definition of the free function was moved at the start of the file
so that it is in scope in all needed places, as there is now no previous
declaration coming from the header.

Change-Id: Ic9a6dadfcfc1f57ba2eb8c8f7e2d45958104b6e3
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>