QDoc: Stop removing duplicate defines
When QDoc parser a project that contains C++ code, it dispatches the
code to Clang to compile it as a mean to extract and sanitize the
user-provided documentation for the project.

When performing this compilation process, QDoc allows to pass
user-provided definitions, to customize the process, to Clang by the use
of the "defines" and "clangdefines" configuration variables.

When QDoc extracts those user-provided defines from the configuration
and passes them to the instances of Clang's compilation when processing
a project, it removes duplicate definitions, and converts the remaining
ones to "-D" arguments for Clang.

Passing multiple of the same defines to Clang is not generally an issue.

While duplicate definitions might indicate an error in the user-provided
configuration, QDoc does not report them, so that they simply get
silently ignored.

Supposedly, the removal is intended to reduce the space that will be
used to store those definitions, generally in a vector, but in practice
this will neither be a common case nor should it be handled at this
level.

Hence, stop removing user-defined duplicate definitions when preparing
the argument for a Clang compilation process.

QDoc performs this process in `ClangCodeParser::initializeParser`, where
a set is used to perform duplicate checkings while the user-provided
definitions are converted into arguments that can be given to Clang.

`ClangCodeParser::initializeParser` is now modified to perform the same
process without performing any duplicate definitions checks.

Change-Id: Ia06528bf8f567b3b639afe83a0e378b941c9a889
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>