QDoc: Remove the clangdefines configuration variable
When QDoc parses a project that contains C++ code, it passes that code
to Clang to extract and sanitize the user provided documentation.

To customize the compilation process that Clang performs, QDoc allows
the user to pass definitions, equivalent to Clang's "-D" argument, to
the compilation process through the use of the "clangdefines" and
"defines" configuration variables.

The two configuration variable have the same purpose but are slightly
different.

In particular, "clangdefines" is not supposed to have "wildcards" in it,
specifically the "*" character, while "defines" are allowed to as they
are used in multiple places in QDoc some of which allow the use of
wildcards.

While this difference could be meaningful, QDoc does not actually ensure
that this is the case.

That is, nothing stops a user from using a wildcard in a "clangdefines"
definition.

Furthermore, when QDoc processes the two variables into "-D" arguments
for Clang, it filters all wildcard-using definitions from the "defines"
configuration variable, making the difference between "defines" and
"clangdefines" less marked.

In turn, it performs no checking for the definitions in the
"clangdefines" configuration variables, so that a user could inject
incorrect definition from the usage of the variable.

The only meaningful difference comes into play when the user wants to
have a definition during the process of the C++ code but not during the
processing of a documentation block itself.

As was mentioned above, "defines" is used in multiple places during a
QDoc process and the user might desire a more narrow scope for specific
definitions.

In pratice, we have seen no such requirements during the many years that
the two configuration variables were available, so that this might be a
very rare usage.

Furthermore, "clangdefines" was never documented so that users of QDoc
will not generally have any knowledge of it, so that in fact zero usages
are present in Qt's current documentation.

Hence, considering the insufficient semantical handling of the
configuration variable on QDoc side, the duplicated feature of the
variable compared to "defines" and the lack of usage in Qt's current
documentation, "clangdefines" is now removed from QDoc.

Each reference to it in the code-base was removed.

The only consumer of the variable, `ClangCodeParser::initializeParser`,
where it was used to prepare the arguments for Clang to compile the
source code, was modified to only extracts the contents of the "defines"
configuration variable, ignoring the contents of the now non-existing
"clangdefines".

Change-Id: I7adbad5889ae74755369530c747673bfbba85bd9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>