QDoc: Make s_fn an instance member
When QDoc processes a project, it parses its sources to extract and
validate the user-provided documentation.

For C++ code specifically, QDoc performs a series of steps.

It first parses the necessary header files to build a PCH, while
producing an internal intermediate representation of the code-base.
Then it uses the PCH while parsing all implementation files and ".qdoc"
files to extract the user provided documentation.

For each documentation block that has an "\fn" topic, it produces a
file, "dummyfile.cpp", that contains the argument of the "\fn" command.
It parses the file as if it was an implementation file and then uses the
produced intermediate representation from the parsing to understand
which element the "\fn" refers to.
This process is interleaved with the parsing of the files that contain
the documentation block.

Most of those processes are performed by `ClangCodeParser`, which
interfaces with Clang's APIs to perform the compilation process and
extract the required information.

When `ClangCodeParser` produces a "dummyfile", it builds its content, in
a way that is suitable to retrieve enough information to match the "\fn"
command with a documentatble entity, and store those contents, before
writing them to a temporary file, in a static member of
`ClangCodeParser`, `s_fn`.

Interfacing with Clang is currently the most expensive operation in
QDoc.
To address this issue and improve the performances of QDoc, it is
intended to parse implementation files in parallel.

In particular, we currently aim to have one `ClangCodeParser` instance
for each source file, performing the parsing process in parallel.

`s_fn`, being a static member that would be shared by all instances,
preludes this parallelization, without synchronization that would defeat
the purpose, as it would be modified/read in a racy way by the instances
when they try to process an "\fn" command as described above.

Most probably, the member was originally static to reduce allocations,
but the effect of it being is shown insignificant by profiling.

Hence, `s_fn` is now an instance member instad of a static member of
`ClangCodeParser`.

The static qualifier was removed from the member declaration in
"clangcodeparser.h".

The initialization of the static member was removed from
`clangcodeparser.cpp`.

`ClangCodeParser::fn`, a static method that acts as a getter for `s_fn`
was remove due to the new instance scope of `s_fn`.

`fn` had a single consumer, the `readFile` free function in
"clangcodeparser.cpp".

When QDoc extracts information from a C++ code-base, it sometimes
extracts that information "as written in the source file".

This is generally performed through the free function `getSpelling` in
"clangcodeparser.cpp", which uses Clang's C APIs to obtain the required
information.

When this is done, Clang will access the original source file,
performing a read.
This was found out to be a bottleneck in certain situations, due to the
very huge amount of repeated reads of the same file.

To avoid the issue, a cache was implemented for recently read files, to
reduce the amount of reads. The content of the cache are hashed based on
the path of a source file, so that source files with the same path are
indistinguishable.

The cache is managed by the `readFile` free function, which dispatches
to an actual read when necessary or to the cached content of a file when
possible.
`readFile` is only called by `getSpelling`, as it is the main produces
of file reads.

As part of this dispatching, `readFile` calls `ClangCodeParser::fn`.
As stated above, QDoc produces a "dummyfile.cpp" when processing an
"\fn" command, and the contents of that will be written in the file are
stored in `s_fn`.

Each "dummyfile.cpp" is indistinguishable from the point of view of the
cache, as the "dummyfile" is reused and there is no unique path to hash.

As, from the point of view of `getSpelling` and `readFile` and due to
the current architecture, it is impossible to discern when the content
of a "dummyfile" changes, `readFile` always re-reads its contents to
obtain the newest information.

To avoid performing reads, it uses `ClangCodeParser::fn` to obtain the
current content of `s_fn`, which is equivalent to the current content of
the "dummyfile".

Nonetheless, `getSpelling` is never used, by the current code-base,
during the implementation file parsing step of `ClangCodeParser` and is
instead used only when parsing header files.

This is due to a change in the architecture of our interface for Clang
APIs that is slowly phasing `getSpelling` out.

Since the "dummyfile" comes into play only when processing an "\fn"
command, which is only done when parsing the implementation files, any
handling for the "dummyfile" in `readFile` is dead-code, as it will
never come into play.

Hence, the special handling of "dummyfile" was removed from `readFile`
and the single usage of `ClangCodeParser::fn` was removed with it.

While this makes it so `getSpelling`/`readFile` have an additional
footgun, as they would imply a cache collision, thus producing incorrect
data, if they were called in a context where a "dummyfile" might be
read, `getSpelling` is not intended to be used anymore for new code and
will be slowly removed when possible, so that we decide to ignore, but
keep an eye open for, this new footgun as we consider it temporary.

Change-Id: I8e5a6fce6cf2877048925127707cd620fcbf1e60
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>