QDoc: Move ClangCodeParser initialization to constructor
When QDoc processes a project, it parses the source code of the project
to extract and sanitize the user-provided documentation.

Internally, a series of classes are used to perform this process
respecting the `CodeParser` interface.

For example, a child class of `CodeParser`, `ClangCodeParser`, takes
care of interfacing with Clang's APIs to parse the C++ code available in
the project.

`CodeParser`s are expected to be single instance objects, as a form or
misguided optimization.

They provide an `initializeParser`/`terminateParser` method pairs that
act as managed constructor/destructor pair to initialize and reset the
internal state of the instances.

Interfacing with Clang was profiled to be the greatest bottleneck in
QDoc. To improve the performance of QDoc and reduce the issue, the
code-base is being changed to parse multiple source files in parallel
when interfacing with Clang,
which is expected to have a strong impact on performance.

As `ClangCodeParser` is the entity responsible for the bottleneck, and
considering a series of architectural issues that makes it difficult to
reform the interfacing part as a whole, the current plan is to have one
`ClangCodeParser` instance per source file, each one parsing a source
file in parallel.

With this objective in mind, `CodeParser` single instance architecture
is unsuitable to what the code-base is moving towards.

While we cannot directly remove the `CodeParser` interface at this
point, we can generally ignore it for `ClangCodeParser` so that we can
move forward.

Hence, a constructor was added to `ClangCodeParser`.
The code originally in `ClangCodeParser::initializeParser` was moved to
this new constructor and `initializeParser` was modified to have an
empty definition.

Some of the code originally in `initializeParser` that handled resetting
the internal state of the instance was removed, as that is not
meaningful anymore with a normal instance-scope.

An debug log in the original `initializeParser` was removed
as it is not meaningful for the current development.

The internal state of `ClangCodeParser` depends on a provided
user-configuration.
QDoc uses a singleton instance for a parsed/non-parsed configuration and
`ClangCodeParser` implicitly accesses this instance when initialiazing.

The configuration instance is available for the whole program but is not
always initialized.

The single existing instance of `ClangCodeParser` in the current
code-base is built before the configuration instance is initialized.

This worked before as the actual initialization was performed by
`initializeParser`, which was called much later down the line when the
configuration was initialized.

Hence, to ensure that the implicitly accessed configuration is
initialized when the single `ClangCodeParser` instance is built, in
"main.cpp", so that the internal state of the instance can be correctly
initialized, the construction of the single `ClangCodeParser` instance
was moved further down the code, at a place where the configuration has
been initialized, to respect the implicit order of operations that is
required by the current architecture.

Change-Id: Id9fe4a19646147685d600b00eea10b437321abc2
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>