QDoc: Make the dependency on Config in ClangCodeParser more explicit
QDoc uses a `Config` class to represent a user-provided
project configuration.

`Config` is built as a singleton that is globally available but it might
or might not be initialized at any point in the program.

Certain other parts of the program depends on `Config` for one reason or
another.

For example, the `ClangCodeParser` depends on `Config` to represent a
parsed project configuration to initialize its state, as part of its
constructor.

This means that constructing an instance of `ClangCodeParser` when the
singleton `Config` does not represent a parsed user-provided project
configuration might result in undefined behavior.

While we cannot completely change the architecture at this point in time
to remove those errors completely, we can at least try to make the
dependency more explicit, so that a developer that is building an
instance of `ClangCodeParser` has to think about the state of `Config`.

Hence, the single constructor for `ClangCodeParser` will now require an
instance of `Config` instead of accessing it directly in the definition.

The declaration of the constructor in "clangcodeparser.h" was modified
to require a const reference to `Config`.
An additional include for "config.h" was added to ensure that the
`Config` type is available as required by the new argument.

The implementation of the constructor in "clangcodeparser.cpp" was
modified to use the new argument instead of obtaining the singleton
`Config` instance directly.

THe single usage of the constructor in "main.cpp" was modified to be
passed the single instance of `Config`.

Change-Id: Ie877e88cd558f0d2dd5dfba2fb9c1f097dc74c7d
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>