QDoc: Add regeneration capabilities to tst_validateQdocOutputFiles
Modifications to QDoc may impact the output it generates. While some
such changes may be highly specific, the introduction of a new command,
for example, other such changes may have a much broader scope. The
test that validates QDoc's output files is vulnerable to such changes,
and doesn't have a convenient way to update the expected content for
all of its various test projects in a consistent manner.

This patch modifies tst_validateQdocOutputFiles such that the test
itself can replace the contents of the `expected` directory for its
`testdata` contents. It does so by checking if an environment variable,
`QDOC_REGENERATE_TESTDATA`, is set. If it is, the test will use the
`expected` directory as output directory for each test project, instead
of generating into a temporary directory. The `expected` directory and
contents is recursively deleted before QDoc runs. After running QDoc,
the test skips the comparison and verification steps. Finally, this
change updates the `README.md` for the test with instruction on how to
regenerate expected content for all tests.

Task-number: QTBUG-119500
Change-Id: Ie06b2b4f485504d2acce11420eed82e1cdf2d221
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>